Move quadtree geometry utils into the utils module system
